+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - (DEV) `npm run test-all-serial` runs all the tests in serial.
 ### Added
 - `gulp build-docs` adds links for property and method classes.
+- Added functions: `log`, `log10` and `exp`.
 
 
 ## [0.3.0] -